webcurl/requirements.txt
flask==3.0.0
requests==2.32.3
pyyaml==6.0.2
tinydb==4.7.0
